requirements.txt
alembic==1.14.0
altgraph==0.17.4
APScheduler==3.10.4
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
msal==1.31.0
packaging==24.2
pefile==2023.2.7
pycparser==2.22
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
PyJWT==2.9.0
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2024.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qbittorrent-api==2024.10.68
redis==5.2.0
requests==2.32.3
setuptools==75.3.0
six==1.16.0
soupsieve==2.6
spotipy==2.24.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
Werkzeug==3.1.2
wfastcgi==3.0.0
WTForms==3.2.1